précision des attaques entre 0 et 1

+++ ressources/creation_tables.sql
@@ -33,7 +33,7 @@ CREATE TABLE move(
 	category varchar(10),
 	pp int,
 	power int,
-	accuracy int
+	accuracy numeric
 );
 
 CREATE TABLE ability(
